% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPOT.R
\name{export}
\alias{export}
\title{Export the optimized pipeline as Python code}
\usage{
export(obj, path)
}
\arguments{
\item{obj}{A \code{TPOTClassifier} or a \code{TPOTRegressor}}

\item{path}{\code{String} containing the path and file name of the desired output file}
}
\value{
The class probabilities of the input samples
}
\description{
Export the optimized pipeline as Python code
}
