% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPOT.R
\name{fit}
\alias{fit}
\title{Fitting a model to the data}
\usage{
fit(obj, features, target, sample_weight = NULL, group = NULL)
}
\arguments{
\item{obj}{A \code{TPOTClassifier} or a \code{TPOTRegressor}}

\item{target}{List of class labels for prediction}

\item{group}{Group labels for the samples used when performing cross-validation.
This parameter should only be used in conjunction with sklearn's Group cross-validation
functions, such as sklearn.model_selection.GroupKFold}

\item{feature}{A \code{data.frame} of observations}

\item{sample_weights}{Per-sample weights. Higher weights indicate more importance. If specified,
sample_weight will be passed to any pipeline element whose fit() function accepts
a sample_weight argument. By default, using sample_weight does not affect tpot's
scoring functions, which determine preferences between pipelines.}
}
\value{
Returns a copy of the fitted TPOT Object
}
\description{
Uses genetic programming to optimize a machine learning pipeline that
maximizes score on the provided features and target. Performs internal
k-fold cross-validaton to avoid overfitting on the provided data. The
best pipeline is then trained on the entire set of provided samples.
}
