% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPOT.R
\name{fit_predict}
\alias{fit_predict}
\title{Call fit and predict in sequence.}
\usage{
fit_predict(obj, features, target, sample_weight = NULL, group = NULL)
}
\arguments{
\item{obj}{A \code{TPOTClassifier} or a \code{TPOTRegressor}}

\item{target}{List of class labels for prediction}

\item{group}{Group labels for the samples used when performing cross-validation.
This parameter should only be used in conjunction with sklearn's Group cross-validation
functions, such as sklearn.model_selection.GroupKFold}
}
\description{
Call fit and predict in sequence.
}
