% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{install_tpot}
\alias{install_tpot}
\title{Install TPOT and its dependencies}
\usage{
install_tpot(method = c("auto", "conda"), conda = "auto",
  version = "default", envname = "r-tpot", extra_packages = NULL,
  extra_pip_packages = NULL, restart_session = TRUE)
}
\arguments{
\item{method}{Installation method. By default, "auto" tries to finds a
method that will work in the local environment. Change the default to force
a specific installation method.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{version}{TPOT version to install. Specify "default" to install
the latest release version.
You can also provide a full major.minor.patch specification (e.g. "1.1.0").}

\item{envname}{Name of Python environment to install within. Default is "r-tpot".}

\item{extra_packages}{Additional Python packages to install along with
TPOT.}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}
}
\description{
Install TPOT and its dependencies
}
