% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPOT.R
\name{predict}
\alias{predict}
\title{Use the optimized pipeline to predict the target for a feature set}
\usage{
predict(obj, ...)
}
\arguments{
\item{obj}{A \code{TPOTClassifier} or a \code{TPOTRegressor}}

\item{feature}{A \code{data.frame} of observations}
}
\value{
Predicted target for the samples in the feature matrix
}
\description{
Use the optimized pipeline to predict the target for a feature set
}
