% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPOT.R
\name{predict_proba}
\alias{predict_proba}
\title{Use the optimized pipeline to estimate the class probabilities for a feature set.}
\usage{
predict_proba(obj, features)
}
\arguments{
\item{obj}{A \code{TPOTClassifier} or a \code{TPOTRegressor}}

\item{feature}{A \code{data.frame} of observations}
}
\value{
The class probabilities of the input samples
}
\description{
Use the optimized pipeline to estimate the class probabilities for a feature set.
}
