% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPOT.R
\name{score}
\alias{score}
\title{Return the score on the given testing data using the user-specified scoring function.}
\usage{
score(obj, testing_features, testing_classes)
}
\arguments{
\item{obj}{A \code{TPOTClassifier} or a \code{TPOTRegressor}}

\item{testing_features}{A \code{data.frame} of the testing set}

\item{testing_classes}{A \code{list} of class labels for prediction in the testing set}
}
\value{
float The estimated test set accuracy
}
\description{
Return the score on the given testing data using the user-specified scoring function.
}
